Added an use section
Currently, the project GVM is called sdkman. After installing GVM I wasn't able to understand how to use it as all documentation was linking to sdkman and obviously non of the commands were working.
With the new `use` section I think it will help people to start using this tool faster.